requirements.txt
requests>=2.31
beautifulsoup4>=4.12
gspread>=5.12
google-auth>=2.25
